Introduce the notion of generators
To enable Property-Based-Testing, we offer a mean to generate values.
For this purpose, we offer the developer generator system.

Because of the syntax, it collides with constructor.
We take the opportunity to remove them and transform them into generators.

This feature impacts the whole system because generators can be split into several files and be imported.
So a refactoring was necessary for the feature to be implemented.

Also, some controls and tests needed to be corrected.

This commit does the following:

* Introduce Generators file syntax and `*.gen` files
* Add parsing of `*.gen` files with `GeneratorsContextParser`
* Add generator syntax in def files
* Add the case in tests.def sample
* Remove the notion of constructor and introduce the notion of generation (=expression targeting a constructor)
* Split `TestsContextParser` to simplify the import system usage
* Introduce `ValidationContext` to control context with external elements as core generators
* Update all controls to use `ValidationContext` instead of `TestsContext`
* Add utility functions in controls package object to simplify reused codes (eg: expression extraction)
* Add controls to validate generators
* Correct failing tests
* Add tests for new feature